requirements.txt
requests==2.31.0
pandas==2.2.0
numpy==1.26.4
scikit-learn==1.4.0
xgboost==2.0.3
mlflow==2.10.2
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
matplotlib==3.8.2
seaborn==0.13.2
scipy==1.12.0
pyyaml==6.0.1
jupyter==1.0.0
nbformat==5.9.2
ipykernel==6.29.2
joblib==1.3.2
